requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
crispy-bootstrap5==0.7
Django==4.2.2
django-bootstrap-modal-forms==3.0.3
django-bootstrap-v5==1.0.11
django-crispy-forms==2.0
django-phonenumber-field==7.1.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
phonenumbers==8.13.14
Pillow==9.5.0
psycopg2==2.9.6
pytz==2023.3
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.6.3
